pyproject.toml
[project]
name = "agent-development-kit-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deprecated>=1.3.1",
    "google-adk[database]==0.3.0",
    "google-generativeai==0.8.5",
    "litellm==1.66.3",
    "psutil==5.9.5",
    "python-dotenv==1.1.0",
    "yfinance==0.2.56",
]
